cashier0001: refactor files and packages names
