+++ pom.xml
@@ -8,11 +8,11 @@
 		<version>2.3.3.RELEASE</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.IT-Ambition</groupId>
-	<artifactId>Version-1</artifactId>
+	<groupId>com.itambition.pos</groupId>
+	<artifactId>casher</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>Version-1</name>
-	<description>Demo project for Spring Boot</description>
+	<name>cashier</name>
+	<description>it ambition cashier system</description>
 
 	<properties>
 		<java.version>1.8</java.version>
