--- src/main/java/com/ITAmbition/Version1/Entity/OrderModel.java
@@ -1,8 +1,0 @@
-package com.ITAmbition.Version1.Entity;
-
-
-import java.util.List;
-
-public class OrderModel {
-    private List<OrderEnitiy> entries;
-}
