--- src/main/java/com/ITAmbition/Version1/Repository/CateogryRepo.java
@@ -1,9 +1,0 @@
-package com.ITAmbition.Version1.Repository;
-
-import com.ITAmbition.Version1.Entity.CateogryEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CateogryRepo extends JpaRepository<CateogryEntity,Integer> {
-}
