+++ src/main/java/com/itambition/pos/controller/InvoiceController.java
@@ -1,16 +1,13 @@
-package com.ITAmbition.Version1.Controller;
+package com.itambition.pos.controller;
 
-import com.ITAmbition.Version1.Service.InvoiceService;
-import com.ITAmbition.Version1.Service.OrderService;
+import com.itambition.pos.service.InvoiceService;
+import com.itambition.pos.service.OrderService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.ModelAndView;
 
 //@RequestMapping(value = "/invoice")
 @RestController
