+++ src/main/java/com/itambition/pos/entity/OrderID.java
@@ -1,4 +1,4 @@
-package com.ITAmbition.Version1.Entity;
+package com.itambition.pos.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,7 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.Embeddable;
-import javax.persistence.Entity;
 import java.io.Serializable;
 
 @Data
